+++ web/svelte.config.js
@@ -1,0 +1,12 @@
+import adapter from '@sveltejs/adapter-auto';
+
+const config = {
+  kit: {
+    adapter,
+    alias: {
+      $lib: 'src/lib'
+    }
+  }
+};
+
+export default config;